moved hooks to navigation
